requirements.txt
pynput
opencv-python
numpy
Pillow
pyperclip